Transfer constants from classes to modules
