+++ framework/modules/common/alert/alert.rb
@@ -2,7 +2,5 @@
 
 # PageObject for Alert
 class Alert < BasePageObject
-  DONT_UPLOAD = 'You can upload only when Wi-Fi is enabled'
-
   text 'snackbar', id: 'snackbar_text'
 end
